pyproject.toml
[tool.poetry]
name = "cookiecutter-poetry"
version = "0.1.0"
description = "Cookiecutter template for a Python package with Poetry."
authors = ["Johan Vergeer <johanvergeer@gmail.com>"]
license = "BSD-3-Clause"
readme = "README.rst"
keywords = ["cookiecutter", "poetry", "template", "package"]
repository = "https://github.com/johanvergeer/cookiecutter-poetry"

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Topic :: Software Development",
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/johanvergeer/cookiecutter-poetry/issues"

[tool.poetry.dependencies]
python = "^3.6"
pyyaml = "^5.3"

[tool.poetry.dev-dependencies]
pytest-cookies = "^0.4.0"
alabaster = "^0.7.12"
black = "^19.10b0"
bump2version = "^1.0.0"
isort = "^4.3.21"
sphinx = "^2.4.4"
sphinx-rtd-theme = "^0.4.3"
watchdog = "^0.10.1"
flake8 = "^3.7.9"

[tool.black]
exclude="{{cookiecutter.project_slug}}"

[tool.isort]
skip="{{cookiecutter.project_slug}}"
line_length=88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
